requirements.txt
Django==2.0.2
django-cors-headers==2.4.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
djangorestframework-gis==0.12
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
pylint==2.1.1
coreapi==2.3.3
